Move pixmap code from util/ to pixmap/
