+++ src/pixmap/loaders/loader_png.c
@@ -8,9 +8,9 @@
 
 #include "taisei.h"
 
-#include "util.h"
 #include "loaders.h"
-#include "../pngcruft.h"
+#include "util.h"
+#include "util/pngcruft.h"
 
 static uchar png_magic[] = { 0x89, 0x50, 0x4E, 0x47, 0x0D, 0x0A, 0x1A, 0x0A };
