+++ src/pixmap/loaders/loader_webp.c
