+++ src/pixmap/loaders/meson.build
