+++ src/pixmap/meson.build
@@ -1,0 +1,7 @@
+
+pixmap_src = files(
+    'pixmap.c',
+)
+
+subdir('loaders')
+pixmap_src += pixmap_loaders_src
