+++ src/util/meson.build
@@ -13,7 +13,6 @@ util_src = files(
     'io.c',
     'kvparser.c',
     'miscmath.c',
-    'pixmap.c',
     'pngcruft.c',
     'rectpack.c',
     'strbuf.c',
@@ -54,6 +53,3 @@ else
     # No have_posix check, it might just work.
     util_src += files('platform_posix.c')
 endif
-
-subdir('pixmap_loaders')
-util_src += pixmap_loaders_src
